Refactoring: join layout and header folder
